[ion] Haptics implementation are on all simulators but Ion::Haptics::isEnabled is per platform.
