+++ ion/src/simulator/android/Makefile
@@ -1,11 +1,12 @@
 ion_src += $(addprefix ion/src/simulator/android/src/cpp/, \
   images.cpp \
-  haptics.cpp \
+  haptics_enabled.cpp \
 )
 
 ion_src += $(addprefix ion/src/simulator/shared/, \
   dummy/callback.cpp \
   dummy/language.cpp \
+  haptics.cpp \
 )
 
 ion_src += ion/src/shared/collect_registers.cpp
