+++ ion/src/simulator/ios/Makefile
@@ -5,7 +5,8 @@ ion_src += $(addprefix ion/src/simulator/ios/, \
 ion_src += $(addprefix ion/src/simulator/shared/, \
   apple/language.m \
   dummy/callback.cpp \
-  dummy/haptics.cpp \
+  dummy/haptics_enabled.cpp \
+  haptics.cpp \
 )
 
 ion_src += ion/src/shared/collect_registers.cpp
