+++ ion/src/simulator/macos/Makefile
@@ -5,9 +5,10 @@ ion_src += $(addprefix ion/src/simulator/macos/, \
 ion_src += $(addprefix ion/src/simulator/shared/, \
   apple/language.m \
   dummy/callback.cpp \
-  dummy/haptics.cpp \
+  dummy/haptics_enabled.cpp \
   collect_registers_x86_64.s \
   collect_registers.cpp \
+  haptics.cpp \
 )
 
 ifeq ($(EPSILON_TELEMETRY),1)
