--- ion/src/simulator/shared/dummy/haptics.cpp
@@ -1,19 +1,0 @@
-#include "../haptics.h"
-
-namespace Ion {
-namespace Simulator {
-namespace Haptics {
-
-void init() {
-}
-
-void shutdown() {
-}
-
-void perform() {
-}
-
-
-}
-}
-}
