+++ ion/src/simulator/shared/dummy/haptics_enabled.cpp
@@ -1,0 +1,15 @@
+#include "../haptics.h"
+
+namespace Ion {
+namespace Simulator {
+namespace Haptics {
+
+bool isEnabled() {
+  /* Dummy default to false as failsafe.
+   * Avoid to get haptics feedback that are not desactivable. */
+  return false;
+}
+
+}
+}
+}
