+++ ion/src/simulator/windows/Makefile
@@ -6,7 +6,8 @@ ion_src += $(addprefix ion/src/simulator/windows/, \
 
 ion_src += $(addprefix ion/src/simulator/shared/, \
   dummy/callback.cpp \
-  dummy/haptics.cpp \
+  dummy/haptics_enabled.cpp \
+  haptics.cpp \
 )
 
 ion_src += ion/src/shared/collect_registers.cpp
